refactor: onboarding 관련 파일 디렉토리 구조 변경
