+++ src/pages/onboarding/InitProcess.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import InitProcessTemplate from '../templates/InitProcessTemplate';
+import InitProcessTemplate from '../../templates/onboarding/InitProcessTemplate';
 
 export default function InitProcess() {
   return (
