+++ src/pages/onboarding/SignIn.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SignInTemplate from '../templates/SignInTemplate';
+import SignInTemplate from '../../templates/onboarding/SignInTemplate';
 
 export default function SignIn() {
   return (
