+++ src/pages/onboarding/SignUp.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SignUpTemplate from '../templates/SignUpTemplate';
+import SignUpTemplate from '../../templates/onboarding/SignUpTemplate';
 
 export default function SignUp() {
   return (
